sellmind/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2
django-celery-beat==2.8.1
django-timezone-field==7.1
djangorestframework==3.16.0
python-dotenv==1.1.0
exceptiongroup==1.2.2
frozenlist==1.6.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.5.3
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
marshmallow==3.26.1
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.77.0
orjson==3.10.18
packaging==24.2
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
python-json-logger==3.3.0
django-cors-headers==4.7.0
faiss-cpu==1.11.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
zstandard==0.23.0
